PLAT-1109 Auto-Complete Inputs: Improve handling of TAB and ENTER
- The auto-complete popup is now automatically opened when the element is focused by either click or TAB.
- The auto-complete popup is now closed when the user presses ENTER while an illegal or empty pattern was entered.
- Tests:
  - We now precisely assert the element that we expect to have the focus, instead of only having a binary focus assertion wrt. the input element.
  - Various simplifications in the base class of entity-based auto-complete tests.
  - Re-organized entity-based auto-complete tests in a reworked set of test classes, for ease of maintenance and understanding.
