+++ platform-ajax/src/test/java/com/softicar/platform/ajax/input/auto/complete/AjaxAutoCompleteEnterKeyTest.java
@@ -31,7 +31,7 @@ public class AjaxAutoCompleteEnterKeyTest extends AbstractAjaxAutoCompleteString
 		waitForServer();
 
 		// assert popup is not closed but input keeps focus
-		assertTrue(isAutoCompletePopupDisplayed());
+		assertFalse(isAutoCompletePopupDisplayed());
 		assertFocused(inputField);
 	}
