--- platform-ajax/src/test/java/com/softicar/platform/ajax/input/auto/complete/entity/AjaxAutoCompleteEntityInputCreatedTest.java
@@ -1,100 +1,0 @@
-package com.softicar.platform.ajax.input.auto.complete.entity;
-
-import com.softicar.platform.dom.elements.input.auto.DomAutoCompleteInput;
-import org.junit.Test;
-
-/**
- * Contains unit tests for {@link DomAutoCompleteInput} interaction phase <b>"1
- * Created"</b> (see {@link AbstractAjaxAutoCompleteEntityTest}).
- *
- * @author Alexander Schmidt
- */
-public class AjaxAutoCompleteEntityInputCreatedTest extends AbstractAjaxAutoCompleteEntityTest {
-
-	@Test
-	public void testValidInput() {
-
-		setup//
-			.setSelectedValue(VALUE1)
-			.execute();
-
-		asserter//
-			.expectValues(VALUE1)
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectNoFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testValidInputAfterChange() {
-
-		setup//
-			.setSelectedValue(VALUE1)
-			.setSelectedValue(VALUE2)
-			.execute();
-
-		asserter//
-			.expectValues(VALUE2)
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectNoFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testIllegalInput() {
-
-		setup//
-			.setSelectedValue(ILLEGAL_VALUE)
-			.execute();
-
-		asserter//
-			.expectInputText(ILLEGAL_VALUE.toDisplayStringWithId())
-			.expectSelectedValueExceptionMessage()
-			.expectIndicatorIllegal()
-			.expectPopupNotDisplayed()
-			.expectNoFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testEmptyInput() {
-
-		setup//
-			.execute();
-
-		asserter//
-			.expectValuesNone()
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectNoFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testEmptyInputAfterClear() {
-
-		setup//
-			.setSelectedValue(VALUE1)
-			.setSelectedValueNone()
-			.execute();
-
-		asserter//
-			.expectValuesNone()
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectNoFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-}
