--- platform-ajax/src/test/java/com/softicar/platform/ajax/input/auto/complete/entity/AjaxAutoCompleteEntityInputFocusedPopupReopenedTest.java
@@ -1,22 +1,0 @@
-package com.softicar.platform.ajax.input.auto.complete.entity;
-
-import com.softicar.platform.dom.elements.input.auto.DomAutoCompleteInput;
-import org.junit.Ignore;
-import org.junit.Test;
-
-/**
- * Contains unit tests for {@link DomAutoCompleteInput} interaction phase
- * <b>"2.4 Popup Re-Opened"</b> (see
- * {@link AbstractAjaxAutoCompleteEntityTest}).
- *
- * @author Alexander Schmidt
- */
-public class AjaxAutoCompleteEntityInputFocusedPopupReopenedTest extends AbstractAjaxAutoCompleteEntityTest {
-
-	@Test
-	@Ignore
-	public void test() {
-
-		// nothing yet
-	}
-}
