--- platform-ajax/src/test/java/com/softicar/platform/ajax/input/auto/complete/entity/AjaxAutoCompleteEntityInputFocusedTest.java
@@ -1,119 +1,0 @@
-package com.softicar.platform.ajax.input.auto.complete.entity;
-
-import com.softicar.platform.dom.elements.input.auto.DomAutoCompleteInput;
-import org.junit.Test;
-
-/**
- * Contains unit tests for {@link DomAutoCompleteInput} interaction phase <b>"2
- * Focused"</b> (see {@link AbstractAjaxAutoCompleteEntityTest}).
- *
- * @author Alexander Schmidt
- */
-public class AjaxAutoCompleteEntityInputFocusedTest extends AbstractAjaxAutoCompleteEntityTest {
-
-	@Test
-	public void testValidInputWithFocusByTab() {
-
-		setup//
-			.setSelectedValue(VALUE1)
-			.execute();
-
-		input//
-			.focusByTab()
-			.waitForServer();
-
-		asserter//
-			.expectValues(VALUE1)
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testIllegalInputWithFocusByClick() {
-
-		setup//
-			.setSelectedValue(ILLEGAL_VALUE)
-			.execute();
-
-		input//
-			.clickInputFieldAndClosePopup()
-			.waitForServer();
-
-		asserter//
-			.expectInputText(ILLEGAL_VALUE.toDisplayStringWithId())
-			.expectSelectedValueExceptionMessage()
-			.expectIndicatorIllegal()
-			.expectPopupNotDisplayed()
-			.expectFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackCountOne()
-			.expectCallbackValue(null)
-			.assertAll();
-	}
-
-	@Test
-	public void testIllegalInputWithFocusByTab() {
-
-		setup//
-			.setSelectedValue(ILLEGAL_VALUE)
-			.execute();
-
-		input//
-			.focusByTab()
-			.waitForServer();
-
-		asserter//
-			.expectInputText(ILLEGAL_VALUE.toDisplayStringWithId())
-			.expectSelectedValueExceptionMessage()
-			.expectIndicatorIllegal()
-			.expectPopupNotDisplayed()
-			.expectFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testEmptyInputWithFocusByClick() {
-
-		setup//
-			.execute();
-
-		input//
-			.clickInputFieldAndClosePopup()
-			.waitForServer();
-
-		asserter//
-			.expectValuesNone()
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-
-	@Test
-	public void testEmptyInputWithFocusByTab() {
-
-		setup//
-			.execute();
-
-		input//
-			.focusByTab()
-			.waitForServer();
-
-		asserter//
-			.expectValuesNone()
-			.expectIndicatorNone()
-			.expectPopupNotDisplayed()
-			.expectFocus()
-			.expectBackdropNotDisplayed()
-			.expectCallbackNone()
-			.assertAll();
-	}
-}
